Moved logic from sidebar to app.tsx
