+++ src/App.tsx
@@ -12,6 +12,43 @@ function App() {
     setSelectedProject(project);
   };
 
+
+  const [projects, setProjects] = useState<Project[]>([]);
+
+  const addProject = (projectName: string) => {
+    const newProject: Project = {
+      id: uuidv4(),
+      projectName,
+      isEditing: false,
+      toDoTasks: [] as ToDoTask[],
+      doingTasks: [] as ToDoTask[],
+      doneTasks: [] as ToDoTask[]
+    };
+    setProjects([...projects, newProject]);
+  };
+
+  const deleteproject = (id: string) => {
+
+    setProjects(projects.filter(project => project.id !== id))
+
+  }
+
+  const startEditingProject = (id: string) => {
+    setProjects(
+      projects.map((project) =>
+        project.id === id ? { ...project, isEditing: true } : project
+      )
+    );
+  };
+
+  const editProjectName = (projectName: string, id: string) => {
+    setProjects(
+      projects.map((p) =>
+        p.id === id? {...p, projectName, isEditing: false } : p
+      )
+    );
+  };
+
   return (
     <>
     <div className="app_container">
