+++ src/types/index.tsx
@@ -1,9 +1,14 @@
 export interface Project {
   id: string;
   projectName: string;
-  completed: boolean;
   isEditing: boolean;
+  toDoTasks: ToDoTask[];
+  doingTasks: ToDoTask[];
+  doneTasks: ToDoTask[];
 }
 
-
-
+export interface ToDoTask {
+  id: string;
+  taskName: string;
+  isEditing: boolean;
+}
